refactor: assets 상수 디렉토리 이동
